requirements.txt
# SAM 2 (Segment Anything 2) Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.5.0

# For downloading and managing model checkpoints
requests>=2.25.0
tqdm>=4.62.0

# For API server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# For image processing
scikit-image>=0.18.0

# SAM 2 specific (will be installed from GitHub)
# git+https://github.com/facebookresearch/segment-anything-2.git 